+++ backuper/main.ps1
@@ -85,6 +85,7 @@ $libsToLoad = @(
     'lib\ui\userdata_edit_dialog.ps1', # Phase 2.7
     'lib\ui\unc_helper.ps1',
     'lib\ui\unc_connect_dialog.ps1',
+    'lib\ui\credentials_select_dialog.ps1', # v0.20.0
     'lib\ui\backup_view.ps1',
     'lib\ui\restore_view.ps1',
     'lib\ui\progress_view.ps1',